requirements/prd.txt
ase==3.23.0
dplutils==0.5.2
pandas==2.2.2
ray[default]==2.24.0
rdkit_pypi==2022.9.5
torch==2.3.1
torch_geometric==2.5.3
